See Description
- Enable Testability in release builds to fix Apple Clang compile bugs that break dynamic_cast
- New: `UIPlanner`, `UIPlan`. Use to build UI models, and interpret them with `ImGuiPlanPainter` to create imGui UIs
- Checkers Example with QuickBuild animations
- Improved Life Example, optimizations
- MatrixBoardAnimation example that shows matrix board animations
- FileProcessorScene (work in progress) for file processors
- Added clang-format to Kaiju project
- Animation improvements: TimeTrack, Timeline, better AnimationCycleTimer, better Valve, Keyframe animations, etc.
- New: `QuickAnimate` for quickly building animations
- Playable improvements. Now supports driver that drives the playable
- New: `TrackFramePlayable` allows you to place animation frames as keyframes on a TimeTrack
- Camera can specify a clear color
- World now has frame rate limiter and stutter fix (breaks large deltas into smaller time steps)
- New: BlinkEffect for blink animations
- Quality of life improvements, more unit tests
- Layouts now build when needed
- New: GridMeshBuilder for rendering grids
- Hover gestures supported via signals/QuickBuild hover
- Drop targets now supported
- Drag handler logic improvements
- _In Progress:_ Working on getting the windows build up and running via Visual Studio. Not quite there yet...

+++ PhaseJump/PhaseJump/Classes/Developer/DevProfiler.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "StringUtils.h"
+#include <chrono>
 
 /*
  RATING: 5 stars
@@ -41,7 +42,7 @@ namespace PJ {
             auto now = std::chrono::steady_clock::now();
             auto duration = now - start;
             double durationInMicroseconds =
-                std::chrono::duration_cast<std::chrono::microseconds>(duration).count();
+                (double)std::chrono::duration_cast<std::chrono::microseconds>(duration).count();
 
             double durationInMilliseconds = 0;
             if (durationInMicroseconds > 0) {
